refactor and outcoming call
