+++ requirements.txt
@@ -5,6 +5,7 @@ aiosignal==1.3.1
 annotated-types==0.7.0
 anyio==4.6.0
 async-timeout==4.0.3
+apscheduler==3.10.4
 attrs==24.2.0
 certifi==2024.8.30
 charset-normalizer==3.3.2
